Refactor memory management with `MemoryHandle`
Replaced `SafeHandle` with a new `MemoryHandle` system for improved memory tracking, safety, and leak detection. Updated allocators (`ArenaAllocator`, `HeapAllocator`, `StackAllocator`) and collections (`UnTypedArray`, `UnsafeArray<T>`, `UnsafeBitSet`) to use `MemoryHandle`.

Refactored `AllocationManager` to use `ConcurrentSlotMap` for live allocation tracking and added methods for managing `MemoryHandle` instances. Simplified alignment and padding logic across allocators and collections.

Enhanced performance with optimized memory operations (`MemClear`, `MemSet`, `MemCpy`) and vectorized operations in `MemoryUtility` and `UnsafeBitSet`. Fixed alignment issues in vectorized memory operations.

Updated tests to reflect the new memory management system and added new tests for `UnsafeBitSet` bitwise operations. Enabled `ENABLE_COLLECTION_CHECKS` for debug builds and improved error messages and documentation.

Removed unused `SafeHandle` code and adjusted project configuration to include necessary references.

+++ Misaki.HighPerformance.LowLevel/Collections/Contracts/IUnsafeCollection.cs
@@ -82,11 +82,11 @@ public interface IUnTypedCollection : IUnsafeCollection
     /// <summary>
     /// The total size of the buffer in bytes.
     /// </summary>
-    uint Size
+    nuint Size
     {
         get;
     }
 
-    ref T GetElementAt<T>(uint index)
+    ref T GetElementAt<T>(nuint index)
         where T : unmanaged;
 }
